requirements.txt
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
# psycopg2-binary is optional (Postgres). Install manually if using PostgreSQL.
# psycopg2-binary==2.9.10
python-dotenv==1.0.0
Werkzeug==2.3.6
